Move Core MIDI stuff around
